move marko builder to common/utils
