+++ packages/common/utils/build-marko-global.js
