+++ packages/inquiry/routers/submit.js
@@ -1,7 +1,7 @@
 const { asyncRoute } = require('@base-cms/utils');
 const { content: contentLoader } = require('@base-cms/web-common/page-loaders');
 const send = require('@endeavorb2b/base-website-common/utils/send-mail');
-const buildMarkoGlobal = require('@endeavorb2b/base-website-routing-utils/build-marko-global');
+const buildMarkoGlobal = require('@endeavorb2b/base-website-common/utils/build-marko-global');
 const { notificationBuilder, confirmationBuilder } = require('../template-builders');
 
 module.exports = ({ queryFragment, notification, confirmation }) => asyncRoute(async (req, res) => {
